Validations and middlewares

+++ Pavas.API/EndPoints/Inventory/Add/AddInventoryRequest.cs
@@ -9,5 +9,5 @@ public record AddInventoryRequest(
     int CompanyId,
     InventoryType Type,
     decimal Price,
-    int Quantity
+    decimal Quantity
 );
